+++ src/scss/App.scss
@@ -12,4 +12,4 @@ $bg-color: #F4F7FB;
 @import './components/Products';
 @import './components/ProductPage';
 @import './components/ShoppingCart';
-
+@import './pages/NotFound';
