+++ src/scss/pages/NotFound.scss
@@ -1,0 +1,43 @@
+.page-notfound {
+	&__wrapper {
+		position: relative;
+		margin: 0 auto;
+		max-width: 1380px;
+	}
+	&__info {
+		position: absolute;
+		top: 15%;
+		left: 10%;
+		max-width: 290px;
+	}
+	&__title {
+		font-weight: 600;
+		font-size: 35px;
+		line-height: 51px;
+		text-transform: uppercase;
+		color: #FFFFFF;
+		padding: 0;
+		margin: 0 0 20px 0;
+	}
+	&__text {
+		font-weight: 400;
+		font-size: 14px;
+		line-height: 20px;
+		color: #FFFFFF;
+		padding: 0;
+		margin: 0 0 20px 0;
+	}
+	&__cta {
+		background: #FFFFFF;
+		border: 1.3px solid #FFFFFF;
+		border-radius: 5px;
+		padding: 15px;
+	}
+	&__subtext {
+		font-weight: 400;
+		font-size: 12px;
+		line-height: 17px;		
+		color: #FFFFFF;		
+		opacity: 0.7;
+	}
+}
